refactor: split init agent to composable/neuri
