+++ src/main.ts
@@ -8,9 +8,9 @@ import { pathfinder as MineflayerPathfinder } from 'mineflayer-pathfinder'
 import { plugin as MineflayerPVP } from 'mineflayer-pvp'
 import { plugin as MineflayerTool } from 'mineflayer-tool'
 
-import { initAgent } from './agents/action/llm'
 import { initBot } from './composables/bot'
 import { botConfig, initEnv } from './composables/config'
+import { initNeuriAgent } from './composables/neuri'
 import { wrapPlugin } from './libs/mineflayer'
 import { LLMAgent } from './plugins/llm-agent'
 import { initLogger } from './utils/logger'
@@ -36,7 +36,7 @@ async function main() {
   const airiClient = new Client({ name: 'minecraft-bot', url: 'ws://localhost:6121/ws' })
 
   // Dynamically load LLMAgent after the bot is initialized
-  const agent = await initAgent(bot)
+  const agent = await initNeuriAgent(bot)
   await bot.loadPlugin(LLMAgent({ agent, airiClient }))
 
   process.on('SIGINT', () => {
